requirements.txt
colpali-engine
pdf2image
GPUtil
accelerate
openai
gradio[mcp]
tqdm
Pillow
gradio-pdf
qdrant-client
numpy
python-dotenv
minio
torch
torchvision
torchaudio